Tests Refactor
Split tests into different files based on test type
Export reusable variables used across the test
